requirements.txt
requests
langchain
langchain-community
python-dotenv
pydantic
duckduckgo-search
wikipedia
openai